requirements.txt
alembic==1.0.0
click==6.7
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
mysqlclient==2.0.1
psycopg2
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.11
Werkzeug==0.14.1
pillow
pytz
